requirements.txt
Django==3.2.3
django-kronos==1.0
paramiko==2.7.2
mysqlclient==2.0.3
python-telegram-bot==13.4.1
telegram==0.0.1
django-environ==0.4.5
pymemcache==3.4.4

elastic-apm==6.2.1